// frontend/src/types/api/conditionOperator.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * mailassist
 * OpenAPI spec version: 0.1.0
 */

/**
 * Logical grouping operator for condition groups.
 */
export type ConditionOperator = typeof ConditionOperator[keyof typeof ConditionOperator];


export const ConditionOperator = {
  AND: 'AND',
  OR: 'OR',
} as const;